Add information about configuration to the help
This also includes a snazzy function to convert the list of themes into
a line broken list, so you don't have to update it with future themes.

I didn't add too much info on custom themes.  The documentation for this
needs to be updated elsewhere, i.e., in the example theme file.