Fix the line number setting auto-updating
The problem was that the change didn't propagate to the global dict as I
thought it would.  So instead, we use the get_state() of the checkbox
(which is backwards by the way).  ui.setup_palette() doesn't seem to be
necessary in this case, so I left it out.  Anyway, it will still be
called when the dialog is closed.