Allow to grow/shrink the height of the variables, stack, and breakpoints lists
For now, you do it by selecting the list you want to grow/shrink and
pressing [ to grow and ] to shrink, but this can easily be changed.
This uses proportional growing/shrinking, just like the vertical -/+
does, i.e., [ multiplies the weight by 1.25 and ] divides the weight by
1.25.  This perhaps should be changed, as it feels a little awkward,
though the "weight" part of it is hard-coded into urwin, and changing
that would not be easy. On the other hand, this isn't too hard to get
used to, and it's pretty easy to get the sizes you want.

All sizes are saved in the configuration file and reloaded in future
uses.  So technically, a picky user could set any weight he wants to
suit his needs.