Add support for custom stringifiers to PuDB
Also included in this commit is a change that makes the custom theme and
custom stringifier preferences persistant, so that the user does not
have to re-enter them if deselects "Custom".  This wasn't easy to split
out, so it's included here.

I still need to add support for manually choosing a the custom
stringifier on a per-variable basis.