Print ellipsis at the end of an overlong unwrapped variable
This was supposed to use the unicode ellipsis (…) when possible, but
this does not work.  The behavior is very strange.  If you replace the
last character with …, it puts a ?.  If you replace the second to last
character, it replaces that *and* the last character with a ?.  If you
replace the third to last character, it works.  Therefore, for now, we
always just use ... (three dots) regardless of the console encoding.