Enable Control-C to pause the debugger
This fixes #42.  Now pressing Control-C while the script is running enters the
debugger in the line that was run (i.e., it does a set_trace()).

Note that this only works if the script was invoked from `pudb` or `python -m
pudb`.  If PuDB was started with `import pudb; pudb.set_trace()`, this does
not work.  But not that even without this patch, Control-C kills Python
entirely in that case (as opposed to entering post-mortem mode).