Don't modify the breakpoints list in-place
The breakpoints list is used internally by bdb, and adding "fake" set_trace
breakpoints to it leads to an inconsistent state inside of bdb where two lists
of breakpoints are not equal to one another, causing it to crash.  Since
set_trace "breakpoints" are not real breakpoints, there is no need for bdb to
know about them.