Ensure var view never splits Unicode chars
TextCanvas wants byte counts, but we were sending column counts. It does
fix the immediate issue, but when a given variable is so long that we
need to scroll, there are now some dangling characters after the final
quote, even when the variable only contains ASCII.

This imports a private variable from urwid.util to avoid depending too
much on apply_target_encoding magic - not sure if a good idea.