Add colors for disabled breakpoints
I have added grayed out colors to all the themes. I'm not familiar with all
the themes (the one I made and use is midnight). I tried to do my best to
choose colors that fit with the theme, but if someone who uses the theme feels
other colors are better, they should fix them.