Add support for custom shells
A custom shell can be added in a similar way to a custom stringifier. Create a
file with a function pudb_shell(_globals, _locals, first_time) at the module
level, and set it in the settings.

The first_time setting is there to match the existing shells, but it seems
useless (a function can easily tell if it has already been called), so it
should perhaps be removed.