Don't return early when "returning" from a module
We still need to run interaction so that the debugger will show the result of
the last line that was run. We just don't want it to treat the end of the
module as a "return".

For instance, if you have a file with a single variable definition

    a = 1

then if you stop on the definition, pudb will show the variable "a: 1" in the
variables view (but it won't show "return: None" after stepping again).

Fixes https://github.com/inducer/pudb/pull/264#issuecomment-326777042.