Fix errant command line expose after search
In PR #353 I added the option to hide the command line window but I
recently noticed that the window errantly appeared if you did a search.
This commit fixes that by resetting the command line window weight to 0
if it should be hidden, after a search is activated. It seems this is
necessary because URWID sometimes rebiases the widths itself and thus a
0 width may get set to a non 0 value and would be exposed.

I also fixed a bug which has always existed, even prior to my first PR
and that is if you opened a search, then did a ctrl+x to the command
window and then again back to the source window then the source window
would get focus instead of the search window. This commit fixes it so
that focus goes back to the search window, if it is still open.

There is also a small change to ui_tools.py is merely to remove a
redundant "else" that I noticed and some misc other tiny aesthetic code
improvements that I noticed in debugger.py.

This PR completely supercedes my previous PR #352 fix for these issues
because PR #352 did not correctly maintain the size of the command
window if it was re-opened.