Add jedi dependency
Fixes https://github.com/inducer/pudb/issues/414

pudb needs jedi for tab completion, but it was not listed as a
dependency. This PR adds the missing dependency.

I'm not sure which version of jedi to depend on. Is there any reason
not to depend on the latest version (currently 0.18)?

I have added the <1 constraint to guard against potential future
breaking changes in jedi