Add a "default" stringifier
This default is essentially what the old "type" stringifier was, but
with the addition that it also handles the logic for how to stringify
the "basic" types. This allows other stringifiers to be used on these
basic types now.

The "type" stringifier is now always just a way of viewing the type() of
a variable.