Display path to current source file dynamically based on terminal width
The path to current source file will now be calculated dynamically. If
there is enough space, the full absolute path will be shown. Otherwise,
the longest possible path will be shown by trimming the full absolute
path from the left.

If there is an update to the terminal width, the path will be
recalculated when the user runs the next line of code, however the method (next, step, continue).
NOTICE: Pressing Ctrl-l will redraw the screen with the current path, it doesn't recalculated the path string.