refactor: replace LGPL code
This contribution replaces code marked as LGPL to avoid conflicts with
the MIT license used in this project. Moreover, it simplifies the code
by removing the *resource parameter in the get_save_config_path
function, given that it isn't used in this project. It also removes
the xdg_data_home global variable, as it isn't in use either.

For context, in Debian we need to resolve this violation to keep the
package in the archive:

  https://bugs.debian.org/1005854