Allow function definition locals() to be modified in the shell
This still requires more testing, especially for the different shells.

For the IPython shell, this requires not making a copy of locals(), but this
also causes the IPython shell to pollute the locals() namespace with all its
internal variables (like _0 and so on), so a better fix is needed.

Fixes #26.