Revert change to IPython shell to make locals editable
This also causes IPython to inject all its magic variables (_, _0,
get_ipython, etc.) into the locals namespace every time it is started. Unless
there is some way to disable this, this is not worth this functionality, since
it pollutes the variables view and also would break any code that happens to
use those same variable names.