Implement jump command (#306)
* Start implementing jump command

Right now it doesn't stop on the line (like t), which would be ideal. I can't
figure out how to make it do that. pdb's j does it, so it should be possible.

It also isn't documented or heavily tested yet.

See issue #129.

* Fix function name that has changed

* Fix the jump command

We need to update the UI manually as end() will run past the current line.
Also, setting a temporary breakpoint is unnecessary.

* Fix keyboard help for jump

* Fix flake8 issues

Co-authored-by: Andreas Klöckner <inform@tiker.net>