Fix startup when no write permissions
Making the parent directories of the config file could cause an exception if the process doesn't have write permissions. This is common with remote debugging in containers. Catch exceptions, log, and continue, same as other parts of the settings that cannot raise exceptions. The next stage of saving the config file will also fail, but this is already caught.