Make the Debugger a singleton and do not reset the stopframe to None during postmortem.
This fixes #607. Making the Debugger a singleton should be enough to fix #607.
HOWEVER, pytest def postmortem() calls .reset() which resets the stopframe.
Somehow, this causes the debugger to stop somewhere inside internal debugger source code,
instead of in the code the postmortem should be done.

So, this also includes an override of the .reset() method and detects when the debugger
is already running and does not reset the stopframe. See also #67.