HACK separate script to test portability of single key presses
Put the NonBufferedConsole class and example calling code into a
separate script, to check the implementation's portability on more
platforms and outside the (more demanding) context of PuDB.

Passed on Windows, Cygwin, WSL, and native Linux. Is assumed to work
on any POSIX supporting Unix, too. Exclusively uses core modules,
does not depend on 3rd party libraries. Worked with Python 3.9, 3.10,
and 3.12. Is believed to not depend on a specific Python version.

This is NOT suitable for upstream. It's a helper for evaluation of
the approach. Not meant for inclusion in the PuDB project.