SQUASH hygiene, single key getter backend detection, enum instead of bools
Prefer a single enum over a set of individual booleans for the detection
of the backend which implements single key getting. These code paths are
strictly alternative and mutually exclusive. Use the catch-all "else"
for the most portable Python input(3) call.

TODO The "nbc" prefix may need adjustment later when renaming the class
to better reflect that it's a single key getter.