SQUASH rename NonBufferedConsole to ConsoleSingleKeyReader
Better reflect the purpose of the platform specific helper. Rename the
class and its get_single_key() method which callers invoke explicitly.

Discuss the purpose of the class and the context when it's used in the
class' doc string for improved awareness. Could have been lost/missed
in the comment.